Refactor View component and add styling
Refactored the View component to include a new CSS file for styling. Added a new CSS class for the view element and applied the w-1/2 utility class. Also added a new CSS class for the h5 element and applied the text-lg and font-semibold utility classes. This improves the visual presentation of the View component.